refactor: merge services to centralize logic
