+++ webtrit_callkeep_android/android/src/main/AndroidManifest.xml
@@ -50,11 +50,6 @@
             android:foregroundServiceType="phoneCall" />
 
         <service
-            android:name=".services.PushNotificationIsolateService"
-            android:exported="false"
-            android:foregroundServiceType="phoneCall" />
-
-        <service
             android:name=".services.ActiveCallService"
             android:exported="false"
             android:foregroundServiceType="phoneCall|microphone|camera" />
